multimedia-qa-app/backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
pydantic
pydantic-settings
pytest
pytest-asyncio
httpx
pymupdf
python-dotenv
